+++ mol_gnn/nn/__init__.py
@@ -1,7 +1,6 @@
+from agg import Aggregation, AggregationRegistry
 from .encoder import GraphEncoder
 from .loss import *
-from .message_passing import (
-    MessagePassing, MultiInputMessagePassing, Aggregation
-)
+from .message_passing import MessagePassing, MultiInputMessagePassing
 from .predictors import *
 from .utils import Activation
