+++ mol_gnn/nn/message_passing/__init__.py
@@ -1,4 +1,4 @@
-from .agg import Aggregation
+from .agg import MessageAggregation
 from .base import MessagePassing
 from .embed import InputEmbedding, OutputEmbedding
 from .impl import ComposableMessagePassing
